Sprint 7/Desafio/Etapas/requirements.txt
requests
boto3
json
os
datetime
pandas
load_dotenv